Split config resource into two separate files
