+++ resources/config.py
@@ -8,7 +8,6 @@ Functions: Set up everything from the config file we use, Load the required conf
 # Imports for ConfigLoader
 import os
 import configparser
-import traceback
 
 def get_config_filename(default_filename):
     """Return the filename; probably not needed at this point."""
@@ -62,7 +61,7 @@ class ConfigLoader():
             ),
         )
         # honestly, this throws an error in the editor but works fine
-        return config.read()
+        return self.parser.read(config)
 
     def load_server_config_setting(self, filename, section, var):
         """Load a specific server config side."""
@@ -115,96 +114,3 @@ class ConfigLoader():
         )
         self.parser.read(loaded_file)
         return str(self.parser.get(section, var))
-
-# We have to import this error since this is a combined file as
-# of now. Eventually move ConfigGenerator to its own file
-from resources.error import ErrorLogging
-
-class ConfigGenerator():
-    """ConfigGenerator"""
-    def __init__(self, bot):
-        self.bot = bot
-        self.server_settings_path = os.path.abspath(
-            os.path.join(
-                os.path.dirname(__file__),
-                '../channel_settings'
-            )
-        )
-
-    async def check_if_config_exists(self, server_id):
-        """Check if a config file exists."""
-        try:
-            if not os.path.exists('%s.ini' % (os.path.join(
-                    self.server_settings_path,
-                    str(server_id)))):
-                return False
-            return True
-        except Exception:
-            await ErrorLogging().log_error(
-                traceback.format_exc(),
-                'ConfigGenerator: checkIfConfigExists'
-            )
-            return True
-
-    async def generate_default_config_file(self, server_id, owner_id):
-        """Generate the config file for a server."""
-        parser = configparser.ConfigParser()
-
-        # Create each section that we need by default; future cogs
-        # may need to handle writing code to modify the config to add sections
-
-        # ServerSettings config['testing'] = {'test': '45', 'test2': 'yes'}
-        parser['ServerSettings'] = {
-            'owner_id': owner_id,
-            'server_id': server_id,
-            'not_accepted_channel_id': 'NOT_SET'
-        }
-
-        # RoleAssignment
-        parser['RoleAssignment'] = {
-            'enabled': False,
-            'role_list': 'NOT_SET',
-            'assignment_channel_id': 'NOT_SET'
-        }
-
-        # JoinPart
-        parser['JoinPart'] = {
-            'member_join_enabled': False,
-            'member_part_enabled': False,
-            'welcome_channel_id': 'NOT_SET',
-            'leave_channel_id': 'NOT_SET'
-        }
-
-        # BettingGame
-        parser['BettingGame'] = {
-            'enabled': False,
-            'bet_channel_id': 'NOT_SET',
-            'minimum_bet': 'NOT_SET'
-        }
-
-        # ApiCommands
-        parser['ApiCommands'] = {
-            'enabled': False,
-            'api_channel_id': 'NOT_SET'
-        }
-
-
-        try:
-            with open('%s.ini' % (
-                os.path.join(
-                    self.server_settings_path,
-                    str(server_id))), 'w'
-                     ) as configfile:
-                parser.write(configfile)
-                return await self.bot.say(
-                    "Configuration file generated. You will need to \
-                    configure the file to your required settings.")
-        except Exception:
-            await ErrorLogging().log_error(
-                traceback.format_exc(),
-                'ConfigGenerator: checkIfConfigExists'
-            )
-            return await self.bot.say(
-                "Error generating configuration file: {0}"
-                .format(traceback.format_exc())
-            )
